Revert "refactor(schema): extract tokenize logic into shared usecase (#11180)"
This reverts commit ef1bb690af2020ac883f1bc1e8bafc53e5f089a8.
